- update to N++ interfaces for 8.4.8 with scintilla 5.3.2 and lexilla 5.2.1
- still open issue: handling of colouralpha and related calls, currently just handled like colour, so alpha info is missing

+++ PythonScript/src/UtfConversion.cpp
@@ -106,7 +106,7 @@ u16string toUtf16(const ConstString<U32> utf32source) {
 }
 
 unsigned int Utf8Decoder::lengthInCharacters() {
-	unsigned int bytesLeftBefore = _bytesLeft;
+	size_t bytesLeftBefore = _bytesLeft;
 	unsigned int byteIndexBefore = _byteIndex;
 	unsigned int length_in_characters = 0;
 	while (!isEnd()) {
